third_party/WebKit/LayoutTests/virtual/gpu/fast/canvas/canvas-webkitLineDash-expected.txt

Basic test for webkitLineDash and webkitLineDashOffset

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL ctx.webkitLineDashOffset should be 0 (of type number). Was undefined (of type undefined).
PASS lineDash[0] is 15
PASS lineDash[1] is 10
PASS ctx.webkitLineDashOffset is 5
PASS ctx.webkitLineDashOffset is 5
PASS successfullyParsed is true

TEST COMPLETE
